+++ ha/Trie/src/ru/spbau/bibaev/homeassignment/Trie.java
@@ -1,4 +1,4 @@
-package ru.spbau.bibaev.ha1;
+package ru.spbau.bibaev.homeassignment;
 
 public interface Trie {
